token splitter update #1
The token splitter functions split tokens by spaces,
tabs, newlines; and by metacharacters even if there are no spaces between
them. They also split tokens by commas, both simple and double. If the
commas are placed in a word they are considered to be quoting the
rest of the word, like in bash.

TODO: Dollars placed between double commas need to be implemented.