token splitter update
The token splitter functions split tokens by spaces,
tabs, newlines; and by metacharacters even if there are no spaces between
them. They also split tokens by quotes, both simple and double. If the
quotes are placed in a word they are considered to be quoting the
rest of the word, like in bash.

TODO: Dollars placed between double quotes need to be implemented.